+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## Version 5.0.1 (Under development)
 
+### AppCenter
+
+* **[Fix]** Fix Concurrent Modification Exception in DefaultChannel on setNetworkRequestsAllowed.
+
 ### App Center Distribute
 
 * **[Fix]** Fix NPE in Distribute module on resume distribute workflow when showing update setup failed dialog.
-* **[Fix]** Fix Concurrent Modification Exception in DefaultChannel on setNetworkRequestsAllowed.
  ___
 
 ## Version 5.0.0